created sidebar and topbar components
