+++ src/App.js
@@ -1,216 +1,181 @@
-import React from "react";
-import "./Home.css";
+import React, {Component} from "react";
+import "./componentStyles/Home.css";
 import image from "./images";
+import Sidebar from "./components/sidebar";
+import Topbar from "./components/topbar";
 
-function App() {
-  return (
-    <div className="page">
-      <div className="sidebar">
-        <div className="logo">
-          <img alt="logo" src={image.logo} />
-        </div>
-        <div className="menu">
-          <ul>
-            <li>
-              <a className="link active" href="/">
-                Home
-              </a>
-            </li>
-            <li>
-              <a className="link" href="/">
-                About
-              </a>
-            </li>
-            <li>
-              <a className="link" href="/">
-                Services
-              </a>
-            </li>
-            <li>
-              <a className="link" href="/">
-                DMCA
-              </a>
-            </li>
-            <li>
-              <a className="link" href="/">
-                Post Book
-              </a>
-            </li>
-          </ul>
-        </div>
-      </div>
-
-      <div className="page-body">
-        <div className="topbar">
-          <div className="search-bar">
-            <input
-              className="search-box"
-              type="text"
-              placeholder="Search Book..."
-            />
-          </div>
-        </div>
-
-        <div className="page-content">
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book1} />
+class App extends Component {
+  render() {
+    return (
+      <div className="page">
+        <Sidebar/>
+  
+        <div className="page-body">
+          <Topbar/>
+  
+          <div className="page-content">
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book1} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="title link">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-            <div className="p-info text-center">
-              <a href="/" className="title link">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book2} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="title link">
+                  The Morgan River The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book2} />
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book3} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="link title">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-            <div className="p-info text-center">
-              <a href="/" className="title link">
-                The Morgan River The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book4} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="link title">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book3} />
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book5} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="link title">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-            <div className="p-info text-center">
-              <a href="/" className="link title">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book6} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="link title">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book4} />
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book7} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="link title">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-            <div className="p-info text-center">
-              <a href="/" className="link title">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book8} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="link title">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book5} />
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book9} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="title link">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-            <div className="p-info text-center">
-              <a href="/" className="link title">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book10} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="title link">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book6} />
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book11} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="title link">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
-            <div className="p-info text-center">
-              <a href="/" className="link title">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
-            </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book7} />
-            </div>
-            <div className="p-info text-center">
-              <a href="/" className="link title">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
-            </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book8} />
-            </div>
-            <div className="p-info text-center">
-              <a href="/" className="link title">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
-            </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book9} />
-            </div>
-            <div className="p-info text-center">
-              <a href="/" className="title link">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
-            </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book10} />
-            </div>
-            <div className="p-info text-center">
-              <a href="/" className="title link">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
-            </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book11} />
-            </div>
-            <div className="p-info text-center">
-              <a href="/" className="title link">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
-            </div>
-          </div>
-
-          <div className="product">
-            <div className="p-image">
-              <img alt="book" src={image.book12} />
-            </div>
-            <div className="p-info text-center">
-              <a href="/" className="title link">
-                The Morgan River
-              </a>
-              <br />
-              <span className="price">$ 200</span>
+  
+            <div className="product">
+              <div className="p-image">
+                <img alt="book" src={image.book12} />
+              </div>
+              <div className="p-info text-center">
+                <a href="/" className="title link">
+                  The Morgan River
+                </a>
+                <br />
+                <span className="price">$ 200</span>
+              </div>
             </div>
           </div>
         </div>
       </div>
-    </div>
-  );
+    );
+
+  }
+  
 }
 
 export default App;
