+++ src/componentStyles/AdminPost.css
